feat: make statusline file name length infinite
this is something that's been bothering me for a while. just want to be
able to see all or most of the file path so i know where the file im
editing is. may need to just make a large number if file paths get too
long, but for the projects im working on, that shouldnt be an issue.
we'll just have to see

Signed-off-by: Matt Provost <mattprovost6@gmail.com>